requirements.txt
psycopg2-binary
validate-docbr